+++ tests/contrib/django/django_app/urls.py
@@ -2,6 +2,7 @@ from django.conf.urls import url
 from django.contrib.auth import login
 from django.contrib.auth.models import User
 from django.http import HttpResponse
+from django.views.generic import TemplateView
 from django.views.decorators.cache import cache_page
 from django.urls import include, path, re_path
 
@@ -49,6 +50,9 @@ urlpatterns = [
     re_path(r"re-path.*/", repath_view),
     path("path/", path_view),
     path("include/", include("tests.contrib.django.django_app.extra_urls")),
+    # This must precede composed-view.
+    url(r"^some-static-view/$", TemplateView.as_view(template_name="my-template.html")),
     url(r"^composed-template-view/$", views.ComposedTemplateView.as_view(), name="composed-template-view"),
     url(r"^composed-get-view/$", views.ComposedGetView.as_view(), name="composed-get-view"),
+    url(r"^composed-view/$", views.ComposedView.as_view(), name="composed-view"),
 ]